Moved FPS code to a mixin
